Update README.md
Fall In Love with PERFUMES FOR MAN AS GIFT
As the festival season is going on, so why not buy some gifts. If you want to give gifts to your fiancé, husband, friend and so on then the perfume is the best gift option. With this you can make their day full of fragrance.  
Choosing the best perfume is a difficult undertaking. These are not simple designations; the difference is in the product's concentration. Many users find the best for them, but it is difficult for them to navigate and thus choose the proper bottle. So, in order to distinguish them and use them wisely, you must first learn more about their composition and function. So, how can you pick the best perfume among Eau de Toilette, Eau De Toilette Spray, Eau De Parfum, and Perfume?
About concentration:
Concentration is a basic mixture of pure raw materials like that fragrances, oils, natural or synthetic materials and so on, which is diluted to facilitate its evaporation. It’s mostly depending on the type of brand.
Finding a new perfume is not the easiest endeavor. There are almost unlimited options, and it can be difficult to know where to begin. Let us be your wingman as we run through some of the best-selling men's perfumes from cosmetics-beauty shop.
1.	Invictus Victory Eau De Parfum Spray
2.	Bad Boy Eau De Toilette Spray
3.	Man Eau De Toilette Spray
4.	Olympea Legend Eau De Parfum Spray
5.	Rochas Man Eau De Toilette Spray
6.	Steel Essence For Man Eau De Toilette Spray
Any of these greatest perfumes for men’s that last a long time should be purchased. Also, regardless of which one you choose, you can be confident that you will not be disappointed because each of them is unique and superior in their own right.